% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_helpers.R
\docType{data}
\name{colors}
\alias{colors}
\alias{white}
\alias{black}
\alias{lightgray_bot}
\alias{gray_bot}
\alias{semidarkgray_bot}
\alias{darkgray_bot}
\alias{lightgray}
\alias{gray}
\alias{darkgray}
\alias{teal}
\alias{burgundy}
\alias{yellow}
\alias{darkblue}
\alias{orange}
\alias{skyblue}
\alias{tan}
\alias{mint}
\alias{bluegray}
\alias{cream}
\alias{lightorange}
\alias{lightteal}
\alias{blue}
\alias{turquoise}
\alias{green}
\alias{yellow_bot}
\alias{darkgreen}
\alias{red}
\alias{lightblue}
\alias{lightyellow}
\alias{lightred}
\alias{blues}
\alias{turquoises}
\alias{greens}
\alias{yellows}
\alias{teals}
\title{Color variables}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 9.
}
\usage{
white

black

lightgray_bot

gray_bot

semidarkgray_bot

darkgray_bot

lightgray

gray

darkgray

teal

burgundy

yellow

darkblue

orange

skyblue

tan

mint

bluegray

cream

lightorange

lightteal

blue

turquoise

green

yellow_bot

darkgreen

red

lightblue

lightyellow

lightred

blues

turquoises

greens

yellows

teals
}
\description{
Variables that conveniently store individual colors or shades of colors. Facilitates standardization
of colors and eliminates the need to keep track of hex values
}
\keyword{datasets}
