% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iecolorsandtheme.R
\docType{data}
\name{ie_palettes}
\alias{ie_palettes}
\title{Custom color palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
ie_palettes
}
\description{
Custom color palettes
}
\keyword{datasets}
