% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_helpers.R
\docType{data}
\name{palettes}
\alias{palettes}
\alias{pal_cat}
\alias{pal_cat2}
\alias{pal_cat2_gray}
\alias{pal_cat3}
\alias{pal_cat4}
\alias{pal_cat5}
\alias{pal_cat6}
\alias{pal_cat7}
\alias{pal_cat8}
\alias{pal_cat9}
\alias{pal_seq}
\alias{pal_seq2}
\alias{pal_seq3}
\alias{pal_seq4}
\alias{pal_seq5}
\alias{pal_seq6}
\alias{pal_seq7}
\alias{pal_seq8}
\alias{pal_seq9}
\alias{pal_div}
\alias{pal_div_gray}
\alias{pal_div2}
\alias{pal_div3}
\alias{pal_div3_gray}
\alias{pal_div4}
\alias{pal_div5}
\alias{pal_div5_gray}
\alias{pal_div6}
\alias{pal_div7}
\alias{pal_div7_gray}
\alias{pal_div8}
\alias{pal_div9}
\alias{pal_div9_gray}
\alias{pal_warn2}
\alias{pal_warn3}
\title{Palettes}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.

An object of class \code{character} of length 4.

An object of class \code{character} of length 5.

An object of class \code{character} of length 6.

An object of class \code{character} of length 7.

An object of class \code{character} of length 8.

An object of class \code{character} of length 9.

An object of class \code{character} of length 9.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.

An object of class \code{character} of length 4.

An object of class \code{character} of length 5.

An object of class \code{character} of length 6.

An object of class \code{character} of length 7.

An object of class \code{character} of length 8.

An object of class \code{character} of length 9.

An object of class \code{character} of length 9.

An object of class \code{character} of length 9.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 4.

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 6.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 8.

An object of class \code{character} of length 9.

An object of class \code{character} of length 9.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.
}
\usage{
pal_cat

pal_cat2

pal_cat2_gray

pal_cat3

pal_cat4

pal_cat5

pal_cat6

pal_cat7

pal_cat8

pal_cat9

pal_seq

pal_seq2

pal_seq3

pal_seq4

pal_seq5

pal_seq6

pal_seq7

pal_seq8

pal_seq9

pal_div

pal_div_gray

pal_div2

pal_div3

pal_div3_gray

pal_div4

pal_div5

pal_div5_gray

pal_div6

pal_div7

pal_div7_gray

pal_div8

pal_div9

pal_div9_gray

pal_warn2

pal_warn3
}
\description{
IE palettes
}
\keyword{datasets}
