% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iecolorsandtheme.R
\name{scale_color_ie}
\alias{scale_color_ie}
\title{Discrete ggplot scale}
\usage{
scale_color_ie(palette = "palette_diverging")
}
\arguments{
\item{palette}{Character. One of "palette_diverging", "palette_sequential", "palette_categorical".}
}
\description{
Discrete ggplot scale
}
