% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iecolorsandtheme.R
\name{scale_fill_ie}
\alias{scale_fill_ie}
\title{Discrete ggplot fill scale}
\usage{
scale_fill_ie(palette = "palette_diverging")
}
\arguments{
\item{palette}{Character. One of "palette_diverging", "palette_sequential", "palette_categorical".}
}
\description{
Discrete ggplot fill scale
}
