% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iecolorsandtheme.R
\name{theme_iedefault}
\alias{theme_iedefault}
\title{Custom ggplot theme}
\usage{
theme_iedefault()
}
\arguments{
\item{base_size}{Numeric. Base font size.}

\item{base_family}{Character. Base font family.}
}
\description{
Custom ggplot theme
}
